Refactor feature extraction, contain the features inside the extractor and simplify the api. This allows the features to be modified by networks so they can communicate more simply.
Decouple the nested actors inside the planner... this is a huge win as you can now add/remove actors as needed and it allows for the fine tuned planner (upcoming) to have a simpler implementation. The planner communicates with the other actors by assigning its outputs to the feature extractor. Downstream actors wil then see the modified features when they execute and extract.